# swagger-codegen/go-server/api/swagger.yaml
openapi: 3.0.1
info:
  title: Store API
  description: Store API documentation
  contact: {}
  license:
    name: unlicensed
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/products/{id}:
    get:
      tags:
      - product-resource
      operationId: getProduct
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
    put:
      tags:
      - product-resource
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
    delete:
      tags:
      - product-resource
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - product-resource
      operationId: partialUpdateProduct
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
...
  /api/products:
    get:
      tags:
      - product-resource
      operationId: getAllProducts
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 20
      - name: sort
        in: query
        description: 'Sorting criteria in the format: property,(asc|desc). Default
          sort order is ascending. Multiple sort criteria are supported.'
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      - name: eagerload
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: boolean
          default: false
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
                x-content-type: '*/*'
    post:
      tags:
      - product-resource
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
...
components:
  schemas:
...
    ProductCategoryDTO:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
      example:
        name: name
        description: description
        id: 1
    ProductDTO:
      required:
      - ean
      - name
      - price
      - productSize
      - sku
      - upc
      type: object
      properties:
        id:
          type: integer
          format: int64
        sku:
          type: string
        upc:
          type: string
        ean:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          minimum: 0
          exclusiveMinimum: false
          type: number
        productSize:
          type: string
          enum:
          - S
          - M
          - L
          - XL
          - XXL
        colors:
          pattern: ^([a-z]+,)*[a-z]+$
          type: string
        image:
          type: array
          items:
            type: string
            format: byte
        imageContentType:
          type: string
        imageSha1:
          maxLength: 40
          minLength: 40
          pattern: '[a-f0-9]{40}'
          type: string
        imageCdnUrl:
          type: string
        thumbnailSha1:
          maxLength: 40
          minLength: 40
          pattern: '[a-f0-9]{40}'
          type: string
        thumbnailCdnUrl:
          type: string
        productCategory:
          $ref: '#/components/schemas/ProductCategoryDTO'
      description: Product sold by the Online store
      example:
        image:
        - ""
        - ""
        imageCdnUrl: imageCdnUrl
        productSize: S
        imageSha1: imageSha1
        upc: upc
        description: description
        colors: colors
        productCategory:
          name: name
          description: description
          id: 1
        ean: ean
        thumbnailCdnUrl: thumbnailCdnUrl
        price: 0.6027456183070403
        imageContentType: imageContentType
        name: name
        thumbnailSha1: thumbnailSha1
        id: 0
        sku: sku
...